custom-elements/instantiating-custom-elements/extensions-to-document-interface/create-element-namespace.html

<!DOCTYPE html>
<html>
<head>
<title>Namespace for createElement is HTML Namespace</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="Namespace for createElement is HTML Namespace">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#extensions-to-document-interface-to-instantiate">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function() {
    var doc = newHTMLDocument();
    var name = 'x-a';
    doc.registerElement(name);
    var customElement = doc.createElement(name);
    
    assert_equals(customElement.namespaceURI, HTML_NAMESPACE, 
            'Wrong custom element namespace');
}, 'Namespace for createElement is HTML Namespace');


test(function() {
    var doc = newHTMLDocument();
    HTML5_ELEMENTS.forEach(function(value) {        
        var obj = doc.createElement(value);        
        var name = 'x-2-' + value;        
        var proto = Object.create(obj.constructor.prototype);        
        doc.registerElement(name, {
            prototype: Object.create(proto),
            extends: value
        });        
        var customElement = doc.createElement(value, name);
            
        assert_equals(customElement.namespaceURI, HTML_NAMESPACE, 
                'Wrong custom element namespace for element extending ' + value);
    });
}, 'Namespace for createElement is HTML Namespace. Type extension specified');

</script>
</body>
</html>
